requirements.txt
altgraph==0.17.4
certifi==2024.8.30
chardet==3.0.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMuPDF==1.24.13
pytesseract==0.3.13
rfc3986==1.5.0
sniffio==1.3.1
